ARTEMIS-4711 Removing the dependcy on Java EE JMS code by XmlDataImporter.
Introducing ConnectionConfigurationAbtract to expose the configuration
to connect to a broker without requiring the JMS code.
